hw updated to requirements
01-gallery - removed unnecesarry properties
02-video - added reset time when video is ended
03-feedback - block submit when fields are ampty

+++ src/js/03-feedback.js
@@ -1,11 +1,12 @@
 const form = document.querySelector('form');
 import throttle from 'lodash.throttle';
+const data = { email: '', message: '' };
 const saveToLocal = () => {
   if (data.message !== '' || data.email !== '') {
     localStorage.setItem('feedback-form-state', JSON.stringify(data));
   }
 };
-const data = {};
+// const data = {};
 const handleInput = event => {
   console.log('pisze');
   const {
@@ -28,12 +29,13 @@ form.addEventListener('input', throttle(saveToLocal, 500));
 
 const handleSubmit = event => {
   event.preventDefault();
-
   const {
     elements: { email, message },
   } = event.currentTarget;
   data.email = email.value;
   data.message = message.value;
+  if (data.email === '' || data.message === '') return alert('fill the data');
+
   console.log(`Email: ${data.email}, Message: ${data.message}`);
   event.currentTarget.reset();
   localStorage.clear();
